# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/application": {
      "get": {
        "operationId": "ApplicationController_allApplications",
        "summary": "Get all applications",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All applications found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      },
      "post": {
        "operationId": "ApplicationController_createAnApplicationWithRandomUUID",
        "summary": "Create an Application with a random UUID",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      }
    },
    "/application/{applicationId}": {
      "get": {
        "operationId": "ApplicationController_getAnApplication",
        "summary": "Get an application by ID",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Application found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      },
      "post": {
        "operationId": "ApplicationController_createAnApplication",
        "summary": "Create an Application with given id",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      },
      "patch": {
        "operationId": "ApplicationController_updateApplication",
        "summary": "Update an existing application",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Application updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      },
      "delete": {
        "operationId": "ApplicationController_deleteApplication",
        "summary": "Delete an application by ID",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hardDelete",
            "required": false,
            "in": "query",
            "description": "Hard delete flag",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Application deleted Successfully!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      }
    },
    "/application/{applicationId}/role": {
      "post": {
        "operationId": "ApplicationController_createRoleWithRandomUUID",
        "summary": "Create a new role for an application with a random uuid",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      }
    },
    "/application/{applicationId}/role/{roleId}": {
      "post": {
        "operationId": "ApplicationController_createRole",
        "summary": "Create a new role for an application with given id",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "description": "Role ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      },
      "delete": {
        "operationId": "ApplicationController_deleteRole",
        "summary": "Delete a role from an application",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "description": "Role ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Role deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      },
      "patch": {
        "operationId": "ApplicationController_updateRole",
        "summary": "Update a role for an application",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "description": "Role ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Role data to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      }
    },
    "/application/{applicationId}/scope": {
      "post": {
        "operationId": "ApplicationController_createScopeWithRandomUUID",
        "summary": "Create a new Scope for an application with random uuid",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScopeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scope created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      }
    },
    "/application/{applicationId}/scope/{scopeId}": {
      "post": {
        "operationId": "ApplicationController_createScope",
        "summary": "Create a new Scope for an application with given id",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scopeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScopeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scope created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      },
      "delete": {
        "operationId": "ApplicationController_deleteScope",
        "summary": "Delete a scope from an application",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scopeId",
            "required": true,
            "in": "path",
            "description": "scope ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "scope deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      },
      "patch": {
        "operationId": "ApplicationController_updateScope",
        "summary": "Update a Scope for an application",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scopeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scopeID",
            "required": true,
            "in": "path",
            "description": "Scope ID",
            "schema": {}
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Scope data to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScopeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scope updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      }
    },
    "/application/{applicationId}/oauth-configuration": {
      "get": {
        "operationId": "ApplicationController_returnOauthConfiguration",
        "summary": "Get OAuth configuration for an application by ID",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Application's configurations are as follows",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Applications"
        ]
      }
    },
...
  },
  "info": {
    "title": "OIDC Protocol setup",
    "description": "This is the project to implement OIDC protocol in nestjs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        },
        "required": [
          "success",
          "message",
          "data"
        ]
      },
...
      "RoleDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Description of the role",
            "example": "Admin role with full permissions"
          },
          "isDefault": {
            "type": "boolean",
            "description": "Indicates if this role is the default role",
            "example": true
          },
          "isSuperRole": {
            "type": "boolean",
            "description": "Indicates if this role is a super role",
            "example": false
          },
          "name": {
            "type": "string",
            "description": "Name of the role",
            "example": "Admin"
          },
          "id": {
            "type": "string",
            "description": "ID of the role (optional)",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        },
        "required": [
          "description",
          "isDefault",
          "isSuperRole",
          "name"
        ]
      },
...
      "CreateApplicationDto": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "description": "Indicates if the application is active",
            "example": true
          },
          "name": {
            "type": "string",
            "description": "Name of the application",
            "example": "My Application"
          },
          "scopes": {
            "description": "List of scopes",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScopeDto"
            }
          },
          "roles": {
            "description": "List of roles",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            }
          },
          "oauthConfiguration": {
            "description": "OAuth configuration",
            "allOf": [
              {
                "$ref": "#/components/schemas/OauthConfiguration"
              }
            ]
          }
        },
        "required": [
          "active",
          "name",
          "scopes",
          "roles",
          "oauthConfiguration"
        ]
      },
...
      "UpdateApplicationDto": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "description": "Indicates if the application is active (optional)",
            "example": true
          },
          "jwtConfiguration": {
            "description": "JWT configuration (optional)",
            "allOf": [
              {
                "$ref": "#/components/schemas/JwtConfiguration"
              }
            ]
          },
          "oauthConfiguration": {
            "description": "OAuth configuration (optional)",
            "allOf": [
              {
                "$ref": "#/components/schemas/OauthConfiguration"
              }
            ]
          },
          "name": {
            "type": "string",
            "description": "Name of the application (optional)",
            "example": "Updated Application Name"
          }
        }
      },
...
    }
  }
}